processor: Add distortion parameters
The distortion class had hard coded parameters by default.
Add parameters that can be set by the user to configure the distortion.
Add set params method and call it from the process block with the
current chain settings.
Add new parameters to the layout.
Add a tone option to the tanh function to scale how hard the clipping is.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>