Delay: Add delay line class
Add a header for the delay class.
Implement delay line class to be used for delay effect.
This will  allow us to delay a certain signal by a number of samples.
This is essentially a circular buffer that matches the size of the sample
buffer block.
This allows us to store all the samples included in the previous block
to be accessed in the current block, only to be overwritten by the
current sample block for the next iteration.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>