delay: Add delay class
Add a delay class that can be used to add a delay effect to the signal
chain.

This uses the delay line to fetch a previous sample from the circular
delay line buffer corresponding to a given delay time and adds that
previous sample to the output sample.
The delayed sample is mixed in depending on the given decay and wetLevel
paramaters.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>