modules: delay: Bypass delay class
Add an option to bypass the delay effect.
Create a private bool bypassed.
Create a method to set bypassed.
Add logic to the delay processing to not add delayed samples to the
output if the module is bypassed.
This will allow it to continue updating the delay line and work
seamlessly when bypass is disabled.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>