modules: distortion: Update wave shaper function with thread safety
When updating the distortion parameters from the editor the standalone
app is crashing.
This seems to be because the editor and GUI is run from a messenger
thread and not the processor thread.
This is possibly a race condition between the threads when the
waveshaper function is updated while the waveshaper is processing.

Check in the setParameters function whether the tone value has changed
and set a lambda variable if the value has changed.
This updated lambda can then be updated in th eprocessor thread and
should be a minimal increase in CPU load.
Add a function to set the new wave shaper function if the tone value has
changed.
This fixes the crash issue and should minimise the amount of latency in
the processor thread.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>