processor: Add update parameters function
Parameters shouldn't be updated in the processor thread as this adds
latency and CPU usage.

Move the parameter updates from the process block to it's own method
that can be called from the editor only when GUI parameters are updated.
Call this function from prepare to play to initialise the default
values.
Call the distortion class's updateWaveShaper method to update the
waveshaper on the processor thread to avoid crashes.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>