editor: Update parameters when values are changed
Parameters should be updated in the GUI thread only when values are
changed.

Add listeners to parameters.
Add parameterValueChanged callback to record when values are changed in
the GUI.
Start timer for timer callback.
Add timer callback to update parameters every timer interval if
parameter values have changed since the last callback.
This will allow debouncing to prevent spamming the updates when slider
inputs are changing.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>