processor: Increase slope of noise gate filter
The noise gate filter was a first order low pass filter and did not have
a steep enough rolloff to effectively cut out noise above the target
frequency.

Use a chain of four second order Butterworth filters to provide 48
db/decade of attenuation.

Lower the minimum of the noise gate to 100 Hz and update the GUI label.
Add a method to create the coefficients and update the filters in the
chain.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>