make, amp: Include impulse response as binary data
Embed an audio file in the application as binary data rather than
relying on an external audio file.
This is more robust as the amp sim class no longer has to search through
directories for the file.
This will also allow the app to be shipped with no external files
needed.

Include thrash_amp.wav as binary data in the make file.
Link the binary data target.

Include BinaryData.h.

Use binary data with the loadImpulseResponse method of the convolution
class instead of using an audio file at runtime.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>