ui: Add a preset panel to manage presets
Create a preset panel UI class with buttons which can be used to call
the preset manager methods.

Add the preset panel to the editor and set the bounds.

Create a preset manager instance in the processor which can be used by
the editor to update parameters in the audio path.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>