ui: Add distortion module panel
Create module panels header to create panels for the audio modules in
the UI.
These submodules will make it easier to position groups of elements in
the plugin editor.

Add the distortion sliders and make them visible as part of this
component.
Add labels to the sliders.
Position sliders and buttons within the panel.
Draw header and outline.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>