service: Update methods to separate builtin and user presets
Add methods to return a list of either user or builtin presets.
Add a method to return the number of builtin presets.
Update getAllPresets() to include factory presets. These must be listed
first as load and delete methods will behave differently for factory
presets at the start of the list.

Don't delete a preset if it is a builtin preset.
Update loadPreset to get the current preset index and only load presets
from file if the index is greater than the number of builtin presets.
If the current preset falls within the builtin presets, load the preset
from the XML files included in binary data.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>