ui: Add submenus to preset list
Set colour of buttons and preset list.
Add submenus to preset list.
Add factory presets to their own submenu and user created presets to a
submenu.
Set item IDs starting with the builtin presets so that user presets can
be differentiated as items with an ID greater than the number of builtin
presets.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>