modules: amp: Adjust mid frequency max gain
Make max gain for the peak filter 0 dB.
This will make the UI more intuitive as the high and low pass
filters are passive and do not have any positive gain.
Now when all the dials are at maximum there will be 0 dB and they
can be lowered to attenuate the different frequency bands.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>