modules: amp: Replace basic three band EQ with basic tonestack
Remove high cut filter and add a high shelf to attenuate high
frequencies more uniformly.
This approximates a tone stack better than rolling off the higher
frequencies with a low pass filter.

Add a low shelf filter to attenuate mid and bass frequencies when the
bass input is turned down.
Also add a boost to this shelf if the treble input is turned down.
This will reduce the attenuation in the low and mid bands when treble is
turned down.
This is done to approximate a Marshall tone stack which boosts low end
frequencies when treble is turned down.
This should allow for better balancing between frequencies.

Set gain of the peak filter between -9 and -20 dbs. Guitar pickups
naturally boost the mid frequencies so the midband should always be
attenuated to balance the frequencies.

Remove initial parameters set for filters in prepare as they will be
updated to the correct coefficients.

Signed-off-by: Jon Friedman <jfriedman97@gmail.com>